rip out airisk, switch to tinymfv-only pipeline
Old pipeline ran a teacher-forced 1-token probe to pick steering sign,
then a guided-CoT eval for the actual delta. The probe sign-check is
broken (CoT thinking only converges after ~20 tokens, so 1-token sign ≠
post-CoT sign), and POS/NEG were drawn from value_labels filtering --
not a real contrastive signal.

New pipeline: extract on persona-branching pairs (POS/NEG share user_msg
+ suffix, differ only in system persona; ported from SSteer), iso-KL
calibrate every method, eval with tinymfv guided CoT at 64 think tokens.
Each stage emits one logger.info demo trace (decoded prompt + gen with
special tokens) for tokenizer/format debugging.

- delete: scripts/airisk_*, scripts/calib/, scripts/traj/,
  src/steering_lite/eval/airisk_dilemmas.py, old smoke tests
- add: data/branching_suffixes.json (550 entries), data/branching.py
  (load_suffixes + make_persona_pairs), scripts/run_tinymfv_sweep.py,
  tests/test_pipeline.py (parametrized over 11 methods, 11 passed in 6.6s)
- modify: attach.train, calibrate.measure_kl, eval/tinymfv: demo traces.
  calibrate: tqdm progbar over secant iters.
  chars.apply: cast to fp32 for cdist (bf16 unsupported).

refs: calibration https://gist.github.com/wassname/6c11cf30b43d8c228bc114795f1019c7
guided https://gist.github.com/wassname/733c568cd29c2a402be4442d6a061899

+++ src/steering_lite/variants/chars.py
@@ -122,8 +122,9 @@ class CHaRS:
         p = state["p"].to(h)
         sigma = float(state["sigma"])
 
-        # k(x, a_i) per token
-        d2    = torch.cdist(h, a) ** 2
+        # k(x, a_i) per token. cdist doesn't support bf16 -- cast just for the
+        # distance, then drop back to h's dtype for the rest of the math.
+        d2    = (torch.cdist(h.float(), a.float()) ** 2).to(h.dtype)
         kern  = torch.exp(-d2 / (2 * sigma ** 2))
         denom = einsum(kern, p, "b s i, i -> b s")
